correção de alguns bugs com as variaveis de ambiente, e testando as macros do dbt

+++ dbt-project/etl_api_finance/models/staging/account/stg_account.sql
@@ -12,7 +12,7 @@ with accounts_raw as (
 accounts as (
 select
     "ClientId"::int as id_cliente, 
-    "Agencia"::int as agencia,
+    "Agencia"::float as agencia,
     "NomeConta"::varchar as tipo_conta, 
     {{ current_timestamp() }} as data_carga
 from accounts_raw
